apiserver: Crude attempt at gRPC interceptors
I wanted to experiment with logging RPC requests. gRPC has interceptors
for doing things in the middle of requests - this is often called
middleware elsewhere.

Only one of each of unary and stream interceptors can be installed for
each server or client, but you can chain them.

This is a really crude interceptor chain which has logging and "metrics"
(which just prints a duration for now). It's not perfect because the
durations repoted by "metrics" and logging are quite different, although
it's in the order of a few 100ns which is probably reasonable but enough
to cause some discrepancies in reporting.

Metric: 1.195283ms
{"level":"info","rpc":"/api.ExampleService/GetUsers","code":"OK","duration":"756.738µs","time":"2019-05-28T15:34:00+01:00"}
Metric: 957.615µs
{"level":"info","rpc":"/api.ExampleService/GetUsers","code":"OK","duration":"662.457µs","time":"2019-05-28T15:34:01+01:00"}
Metric: 765.284µs
{"level":"info","rpc":"/api.ExampleService/GetUsers","code":"OK","duration":"527.642µs","time":"2019-05-28T15:34:01+01:00"}